Refactor Slider Widget with merging class
- Merge d-slider-track-selected into d-slider-track
- Merge d-slider-range into d-slider
- Merge d-slider-min, d-slider-max into d-slider-label
