+++ src/main/typescript/wcardinal/ui/d-slider-horizontal.ts
@@ -3,11 +3,13 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
+import { Point } from "pixi.js";
 import { DSlider, DSliderOptions, DThemeSlider } from "./d-slider";
-import { DSliderMax } from "./d-slider-max";
-import { DSliderMin } from "./d-slider-min";
-import { DSliderRange } from "./d-slider-range";
-import { DSliderRangeHorizontal } from "./d-slider-range-horizontal";
+import { DSliderLabel } from "./d-slider-label";
+import { DSliderThumb } from "./d-slider-thumb";
+import { DSliderTrack } from "./d-slider-track";
+import { DSliderTrackHorizontal } from "./d-slider-track-horizontal";
+import { DSliderValue } from "./d-slider-value";
 
 export interface DSliderHorizontalOptions<THEME extends DThemeSliderHorizontal> extends DSliderOptions<THEME> {
 
@@ -25,26 +27,86 @@ export class DSliderHorizontal<
 	OPTIONS extends DSliderHorizontalOptions<THEME> = DSliderHorizontalOptions<THEME>
 > extends DSlider<THEME, OPTIONS> {
 
-	protected createRange(): DSliderRange {
-		return new DSliderRangeHorizontal({
+	protected createTrack(): DSliderTrack {
+		return new DSliderTrackHorizontal();
+	}
+
+	protected createValue(): DSliderValue {
+		return new DSliderValue({
 			y: 0
 		});
 	}
 
-	protected createMax(): DSliderMax {
-		return new DSliderMax({});
-	}
-
-	protected createMin(): DSliderMin {
-		return new DSliderMin({
-			x: 0
+	protected createThumb(): DSliderThumb {
+		return new DSliderThumb({
+			x: 0,
+			y: this._value.height + HORIZONTAL_PIXEL_BALANCE
 		});
 	}
 
+	protected createMax(): DSliderLabel {
+		return new DSliderLabel({value: 1});
+	}
+
+	protected createMin(): DSliderLabel {
+		return new DSliderLabel({
+			x: 0,
+			value: 0
+		});
+	}
+
+	protected createTrackSelected(): DSliderTrack {
+		return new DSliderTrackHorizontal();
+	}
+
 	protected updateCoordinate(): void {
-		this._min.y = this._range.yOffset - (this._min.height / 2 - this._range.track.height / 2);
-		this._max.x = this._min.width + this._range.width + HORIZONTAL_PIXEL_BALANCE;
-		this._max.y = this._range.yOffset - (this._min.height / 2 - this._range.track.height / 2);
+		this._trackSelected.width = 0;
+
+		// calculate y-offset to determine y-coordinate of slider bar
+		this._yOffset = this._value.height + HORIZONTAL_PIXEL_BALANCE +
+						this._thumb.height / 2 - this._track.height / 2;
+		this._track.y = this._yOffset;
+		this._trackSelected.y = this._yOffset;
+
+		// this is offset beetween x-coordinate of slider value and slider button
+		this._offset = this._value.width / 2 - this._thumb.width / 2;
+		this._value.x = this._thumb.x - this._offset;
+
+		this._min.x = 0 - this._min.width;
+		this._min.y = this.yOffset - (this._min.height / 2 - this.track.height / 2);
+		this._max.x = this.width + HORIZONTAL_PIXEL_BALANCE;
+		this._max.y = this.yOffset - (this._min.height / 2 - this.track.height / 2);
+	}
+
+	protected onPick( global: Point ) {
+		const point = new Point(0, 0);
+		this.toLocal( global, undefined, point );
+		const x = Math.max( 0, Math.min( this._track.width, point.x ) );
+		if ( x < this._thumb.width / 2 ) {
+			this._thumb.x = 0;
+		} else if (x > (this._track.width - this._thumb.width / 2)) {
+			this._thumb.x = this._track.width - this._thumb.width;
+		} else {
+			this._thumb.x = x - this._thumb.width / 2;
+		}
+		this._ratioValue = x / this._track.width;
+		this._trackSelected.width = this._thumb.x;
+		this._value.x = this._thumb.x - this._offset;
+		this.updateValue(this._min.value, this._max.value);
+	}
+
+	updateThumb(min: number, max: number, value: number) {
+		this._ratioValue = (value - min) / (max - min);
+		const x =  this._ratioValue * this._track.width;
+		if (x > (this._track.width - this._thumb.width / 2)) {
+			this._thumb.x = this._track.width - this._thumb.width;
+		} else {
+			this._thumb.x = x;
+		}
+		this._trackSelected.width = this._thumb.x;
+		this._value.x = this._thumb.x - this._offset;
+		this._value.value = value;
+		this._value.text = String(value);
 	}
 
 	protected getType(): string {
