--- src/main/typescript/wcardinal/ui/theme/white/d-theme-white-slider-range-vertical.ts
@@ -1,19 +1,0 @@
-/*
- * Copyright (C) 2019 Toshiba Corporation
- * SPDX-License-Identifier: Apache-2.0
- */
-
-import { DCoordinateSize } from "../../d-coordinate";
-import { DThemeWhiteSliderRange } from "./d-theme-white-slider-range";
-
-export class DThemeWhiteSliderRangeVertical extends DThemeWhiteSliderRange implements DThemeWhiteSliderRangeVertical {
-
-	getHeight(): DCoordinateSize {
-		return 322;
-	}
-
-	getWidth(): DCoordinateSize {
-		return 35;
-	}
-
-}
