--- src/main/typescript/wcardinal/ui/theme/white/d-theme-white-slider-track-selected.ts
@@ -1,23 +1,0 @@
-/*
- * Copyright (C) 2019 Toshiba Corporation
- * SPDX-License-Identifier: Apache-2.0
- */
-
-import { DBaseState } from "../../d-base-state";
-import { DThemeWhiteButton } from "./d-theme-white-button";
-
-export class DThemeWhiteSliderTrackSelected extends DThemeWhiteButton implements DThemeWhiteSliderTrackSelected {
-
-	getBackgroundColor( state: DBaseState ): number | null {
-		return 0x3399FF;
-	}
-
-	getBorderColor( state: DBaseState ): number | null {
-		return null;
-	}
-
-	getColor( state: DBaseState ): number {
-		return 0x5F5F5F;
-	}
-
-}
