+++ src/main/typescript/wcardinal/ui/theme/white/d-theme-white-slider-vertical.ts
@@ -3,8 +3,15 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
+import { DCoordinateSize } from "../../d-coordinate";
 import { DThemeWhiteSlider } from "./d-theme-white-slider";
 
 export class DThemeWhiteSliderVertical extends DThemeWhiteSlider implements DThemeWhiteSliderVertical {
+	getHeight(): DCoordinateSize {
+		return 322;
+	}
 
+	getWidth(): DCoordinateSize {
+		return 35;
+	}
 }
