+++ src/main/typescript/wcardinal/ui/theme/white/d-theme-white-slider.ts
@@ -4,18 +4,9 @@
  */
 
 import { DBaseInteractive } from "../../d-base-interactive";
-import { DCoordinateSize } from "../../d-coordinate";
 import { DThemeWhiteBase } from "./d-theme-white-base";
 
 export class DThemeWhiteSlider extends DThemeWhiteBase implements DThemeWhiteSlider {
-
-	getHeight(): DCoordinateSize {
-		return "auto";
-	}
-
-	getWidth(): DCoordinateSize {
-		return "auto";
-	}
 
 	getInteractive(): DBaseInteractive {
 		return DBaseInteractive.CHILDREN;
