+++ src/test/resources/templates/slider.html
@@ -16,7 +16,7 @@
 
 		const application = new wcardinal.ui.DApplication();
 
-		new wcardinal.ui.DSlider({
+		new wcardinal.ui.DSliderHorizontal({
 			parent: application.stage,
 			x: 100, y: 100
 		});
